Refactor: Separate calendar parsing logic into utility module
